Moving ES-files to storage module and others back.
