+++ graylog2-server/src/test/java/org/graylog/plugins/views/search/elasticsearch/IndexLookupTest.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog.storage.elasticsearch6.views;
+package org.graylog.plugins.views.search.elasticsearch;
 
 import org.graylog.plugins.views.search.elasticsearch.IndexLookup;
 import org.graylog2.indexer.ranges.IndexRange;
